refactor: comment interface, type 폴더로 분리
