+++ src/pages/ProjectDetail/type/comment.ts
@@ -1,0 +1,10 @@
+interface CommentData {
+  from: string;
+  date: string;
+  content: string;
+}
+
+export interface CommnetBoxProps {
+  commentData: CommentData;
+  lastElement: boolean;
+}
